Create sequence_alignment
This part is to align the two sequences in both DNA level and translated protein level using the same DNA sequence, it will report the alignment results and the alignment scores (for both alignments).  And then I generate 1,000 random sequences to do the sequence alignment and collect the alignment scores to build a statistical model and exam the significance of alignment results. Eventually, I compare the significance of both alignment results.