sequence_alignment
This part is to align the two sequences using both DNA sequences and translated protein sequences, report the alignment results and the alignment scores (for both alignments).  And then generate 1,000 random sequences to do the sequence alignment and collect the alignment scores to build a statistical model and exam the significance of alignment results. Eventually, compare the significance of both alignment results.