simple docking simulation
extract the taxol from PDB file and randomly put it into a virtual cubic box whose side length is 100 Angstrom. Then put the taxol into a random location and make it translate (0~10√3 Angstrom toward random direction) and rotate (0~60 degree around 3 axis) every step for 50 steps.  When the taxol (geometry center) goes out of the box, please move it back by letting it either bounce back or come back from the other side of the box (e.g. when it go out from the top of the box, let it come back from the bottom of the box).